requirements.txt
# Core Deep Learning
torch>=2.0.0
torchvision>=0.15.0
ultralytics>=8.0.0

# Computer Vision
opencv-python>=4.8.0
opencv-contrib-python>=4.8.0
Pillow>=10.0.0
albumentations>=1.3.0

# Data Processing
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.10.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.14.0

# Configuration
pyyaml>=6.0
python-dotenv>=1.0.0

# Progress & Logging
tqdm>=4.65.0
tensorboard>=2.13.0

# API Development
flask>=2.3.0
fastapi>=0.100.0
uvicorn>=0.23.0
python-multipart>=0.0.6

# Model Export & Optimization
onnx>=1.14.0
onnxruntime>=1.15.0
onnxruntime-gpu>=1.15.0  # Optional: GPU support

# Optional: TensorRT (NVIDIA GPUs)
# nvidia-tensorrt>=8.6.0

# Video Processing
imageio>=2.31.0
imageio-ffmpeg>=0.4.8

# Utilities
requests>=2.31.0
scikit-learn>=1.3.0
scikit-image>=0.21.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Code Quality
black>=23.0.0
flake8>=6.0.0
isort>=5.12.0

# Jupyter
jupyter>=1.0.0
notebook>=7.0.0
ipywidgets>=8.0.0

# Object Tracking (Optional)
# deep-sort-realtime>=1.3.2
# lap>=0.4.0

# Additional utilities
imgaug>=0.4.0
pycocotools>=2.0.0
